requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
mysqlclient
python-decouple
cloudinary
django-cloudinary-storage
redis
